noti/requirements.txt
selenium==4.5.0
python-telegram-bot==13.14
schedule==1.1.0